Add option on map mode to support looking for files in different places
The GCS_HELPER_MAP_EXTRA_PREFIXES option will provide a list of extra
folders where gcs-helper should look for files when running in mapped
mode.

The main goal of this change is to be able to support video files and
caption files in different "folders". This can also be expanded to
change the way video files are organized.