Simplify mapping
Simplifying here means one thing: removing a lot of the stuff that we
added to be able to handle some of NYT's business logic, we'll move it
into a closed source piece. To support that, the actual mapping now
happens in a library that can be used independently from the service
itself. The service still works just as fine, but it's waaay simpler.

Mapping in gcs-helper only needs two things: a prefix and a regular
expression to filter the objects that match the filter.

Extra resources, additional prefixes and multiple regexp filters should
not be part of the core of gcs-helper.